Add files via upload
A Library Management System is essential for any institution that maintains a collection of books. 
Traditionally, libraries have relied on manual record-keeping, which can be time-consuming and error-prone. 
In this project, the system is implemented using Linux shell scripting, offering a text-based, command-line interface 
that efficiently performs tasks like adding, searching, viewing, and deleting book records